Add horizontal lines and their legend labels to a plot

Add horizontal lines and their legend labels to a plot

add_horizontal_lines(
  plot,
  plot_data,
  agg_label = NULL,
  color_comb = NULL,
  new_hline_col = NULL,
  new_vline_col = NULL,
  hline_arb = NULL,
  hline_arb_color = "red",
  hline_arb_label = NULL,
  hline_vars = NULL,
  hline_vars_colors = NULL,
  hline_vars_labels = NULL,
  vline_arb = NULL,
  vline_arb_color = "green",
  vline_arb_label = NULL,
  vline_vars = NULL,
  vline_vars_colors = NULL,
  vline_vars_labels = NULL
)

Arguments

plot
the ggplot2 plot object which the horizontal lines should be added to

plot_data
data frame with variables which will be displayed in the plot.

agg_label
label for the line denoting the Mean or Median.

color_comb
character denoting the color of the Mean or Median line.

new_hline_col
the name of the column to be added to plot_data that will hold the single value for the arbitrary horizontal line

new_vline_col
the name of the column to be added to plot_data that will hold the single value for the arbitrary vertical line

hline_arb
numeric value identifying intercept for arbitrary horizontal line.

hline_arb_color
color for the arbitrary horizontal line that will appear on the plot.

hline_arb_label
label for the arbitrary horizontal line that will appear on the legend.

hline_vars
name(s) of variables `(ANR*)` or values `(*LOQ)` identifying intercept values.

hline_vars_colors
color(s) for the horizontal lines defined by variables that will appear on the plot.

hline_vars_labels
labels(s) for the horizontal lines defined by variables that will appear on the legend.

vline_arb
numeric value identifying intercept for arbitrary vertical line.

vline_arb_color
color for the arbitrary vertical line that will appear on the plot.

vline_arb_label
label for the arbitrary vertical that will appear on the legend.

vline_vars
name(s) of variables `(ANR*)` or values `(*LOQ)` identifying intercept values.

vline_vars_colors
color(s) for the vertical lines defined by variables that will appear on the plot.

vline_vars_labels
labels(s) for the vertical lines defined by variables that will appear on the legend.