Function to create a boxplot.

A box plot is a method for graphically depicting groups of numerical data through their quartiles. Box plots may also have lines extending vertically from the boxes (whiskers) indicating variability outside the upper and lower quartiles, hence the term box-and-whisker. Outliers may be plotted as individual points. Box plots are non-parametric: they display variation in samples of a statistical population without making any assumptions of the underlying statistical distribution. The spacings between the different parts of the box indicate the degree of dispersion (spread) and skewness in the data, and show outliers. In addition to the points themselves, they allow one to visually estimate various L-estimators, notably the interquartile range, midhinge, range, mid-range, and trimean.

g_boxplot(data, biomarker, yaxis_var, trt_group, xaxis_var = NULL,
  loq_flag = "LOQFL", unit = NULL, color_manual = NULL,
  shape_manual = NULL, box = TRUE, ymax_scale = NULL,
  ymin_scale = NULL, dot_size = 2, alpha = 1, facet_ncol = NULL,
  hline = NULL, rotate_xlab = FALSE, font_size = NULL,
  armlabel = NULL, facet = NULL)

Arguments

data
data frame with variables which will be displayed in the plot.

biomarker
biomarker to visualize e.g. IGG.

yaxis_var
name of variable containing biomarker results displayed on Y-axis e.g. AVAL.

trt_group
name of variable representing treatment trt_group e.g. ARM.

xaxis_var
variable used to group the data on the x-axis.

loq_flag
name of variable containing LOQ flag e.g. LOQFL.

unit
biomarker unit label e.g. (U/L)

color_manual
vector of colour for trt_group

shape_manual
vector of shapes (used with log_flag)

box
add boxes to the plot (boolean)

ymax_scale
maximum value for the Y axis

ymin_scale
minimum value for the Y axis

dot_size
plot dot size.

alpha
dot transparency (0 = transparent, 1 = opaque)

facet_ncol
number of facets per row. NULL = Use the default for facet_wrap

hline
y-axis value to position a horizontal line. NULL = No line

rotate_xlab
45 degree rotation of x-axis label values.

font_size
point size of tex to use. NULL is use default size

armlabel
label for the treatment symbols in the legend. If not specified then the label attribute for trt_group will be used. If there is no label attribute for trt_group, then the name of the parameter (in title case) will be used.

facet
variable to facet the plot by, or "None" if no faceting required.

Value

ggplot object

Examples

# NOT RUN {
# Example using ADaM structure analysis dataset.

library(dplyr)
library(ggplot2)
library(goshawk)
library(random.cdisc.data)
ADSL <- radsl(N = 100, seed = 1)
ADLB <- radlb(ADSL, seed = 2)
ADLB <- ADLB %>% subset(PARAMCD == "CRP")

g_boxplot(ADLB,
          biomarker = "CRP",
          yaxis_var = "AVAL",
          trt_group = "ARM",
          loq_flag = "LOQFL",
          unit = "AVALU",
          shape_manual = c('N' = 1, 'Y' = 2, 'NA' = NULL),
          hline = NULL,
          facet = "AVISIT",
          xaxis_var = "STUDYID",
          alpha = 0.5,
          rotate_xlab = TRUE
)

# }