Function to create a scatter plot.

Default plot displays scatter facetted by visit with color attributed treatment arms and symbol attributed LOQ values.

g_scatterplot(label = "Scatter Plot", data, param_var = "PARAMCD",
  param = "CRP", xaxis_var = "BASE", yaxis_var = "AVAL",
  trt_group = "ARM", visit = "AVISITCD", loq_flag_var = "LOQFL",
  unit = "AVALU", xmin = NA, xmax = NA, ymin = NA, ymax = NA,
  color_manual = NULL, shape_manual = NULL, facet_ncol = 2,
  facet = FALSE, facet_var = "ARM", reg_line = FALSE, hline = NULL,
  rotate_xlab = FALSE, font_size = 12, dot_size = NULL,
  reg_text_size = 3)

Arguments

label
text string to used to identify plot.

data
ADaM structured analysis laboratory data frame e.g. ALB.

param_var
name of variable containing biomarker codes e.g. PARAMCD.

param
biomarker to visualize e.g. IGG.

xaxis_var
name of variable containing biomarker results displayed on X-axis e.g. BASE.

yaxis_var
name of variable containing biomarker results displayed on Y-axise.g. AVAL.

trt_group
name of variable representing treatment group e.g. ARM.

visit
name of variable containing nominal visits e.g. AVISITCD.

loq_flag_var
name of variable containing LOQ flag e.g. LOQFL.

unit
name of variable containing biomarker unit e.g. AVALU.

xmin
x-axis lower zoom limit.

xmax
x-axis upper zoom limit.

ymin
y-axis lower zoom limit.

ymax
y-axis upper zoom limit.

color_manual
vector of colors applied to treatment values.

shape_manual
vector of symbols applied to LOQ values.

facet_ncol
number of facets per row.

facet
set layout to use treatment facetting.

facet_var
variable to use for facetting beyond visit.

reg_line
include regression line and annotations for slope and coefficient in visualization. Use with facet = TRUE.

hline
y-axis value to position a horizontal line.

rotate_xlab
45 degree rotation of x-axis label values.

font_size
font size control for title, x-axis label, y-axis label and legend.

dot_size
plot dot size.

reg_text_size
font size control for regression line annotations.

Details

Regression uses deming model.

Examples

# NOT RUN {
# Example using ADaM structure analysis dataset.
# ALB refers to biomarker data stored in expected laboratory structure.

param <- c('ADIGG') # FOR TESTING: woud come from teal.goshawk.tm_g_scatterplot.R

plot1 <- g_scatterplot(label = 'Scatter Plot',
           data = ALB,
           param_var = 'PARAMCD',
           param = param,
           xaxis_var = 'BASE',
           yaxis_var = 'AVAL',
           trt_group = 'ARM',
           visit = 'AVISITCD',
           loq_flag_var = 'LOQFL',
           unit = 'AVALU',
           xmin = 0,
           xmax = 200,
           ymin = 0,
           ymax = 200,
           color_manual = color_manual,
           shape_manual = shape_manual,
           facet_ncol = 4,
           facet = FALSE,
           facet_var = "ARM",
           reg_line = FALSE,
           hline = NULL,
           rotate_xlab = FALSE,
           font_size = 14,
           dot_size = 2,
           reg_text_size = 3)
plot1

# }