Function to create a table of descriptive summary statistics to accompany plots.

Output descriptive summary statistics table as a data frame. Includes biomarker, treatment, visit, n, mean, median, sd, min, max,

t_summarytable(
  data,
  trt_group,
  param_var,
  param,
  xaxis_var,
  facet_var = "AVISITCD",
  loq_flag_var = "LOQFL",
  ...
)

Arguments

data
name of data frame to summarize.

trt_group
treatment group variable name e.g. ARM.

param_var
name of variable containing biomarker codes e.g. PARAMCD.

param
biomarker to visualize e.g. IGG.

xaxis_var
name of variable containing biomarker results displayed on X-axis e.g. AVAL.

facet_var
name of variable facetted on typically containing visit values e.g. AVISITCD. If NULL then ignored.

loq_flag_var
name of variable containing LOQ flag e.g. LOQFL.

...
additional options

Details

provide additional information as needed. link to specification file http://rstudio.com

Author

Nick Paszty (npaszty) paszty.nicholas@gene.com

Balazs Toth (tothb2) toth.balazs@gene.com

Examples

# Example using ADaM structure analysis dataset.

library(scda)
library(stringr)

# original ARM value = dose value
arm_mapping <- list("A: Drug X" = "150mg QD", "B: Placebo" = "Placebo",
"C: Combination" = "Combination")

ASL <- synthetic_cdisc_data("latest")$adsl
ADLB <- synthetic_cdisc_data("latest")$adlb
ADLB <- ADLB %>%
  mutate(AVISITCD = case_when(
    AVISIT == "SCREENING" ~ "SCR",
    AVISIT == "BASELINE" ~ "BL",
    grepl("WEEK", AVISIT) ~
      paste(
        "W",
        trimws(
          substr(
            AVISIT,
            start = 6,
            stop = str_locate(AVISIT, "DAY") - 1
          )
        )
      ),
    TRUE ~ NA_character_)) %>%
  mutate(AVISITCDN = case_when(
    AVISITCD == "SCR" ~ -2,
    AVISITCD == "BL" ~ 0,
    grepl("W", AVISITCD) ~ as.numeric(gsub("\\D+", "", AVISITCD)),
    TRUE ~ NA_real_)) %>%
  # use ARMCD values to order treatment in visualization legend
  mutate(TRTORD = ifelse(grepl("C", ARMCD), 1,
    ifelse(grepl("B", ARMCD), 2,
      ifelse(grepl("A", ARMCD), 3, NA)))) %>%
  mutate(ARM = as.character(arm_mapping[match(ARM, names(arm_mapping))])) %>%
  mutate(ARM = factor(ARM) %>%
  reorder(TRTORD))

tbl <- t_summarytable(data = ADLB,
               trt_group = "ARM",
               param_var = "PARAMCD",
               param = c("CRP"),
               xaxis_var = "AVAL",
               facet_var = "AVISITCD",
               loq_flag_var = "LOQFL")
tbl
#>    StudyID AnlVar Biomarker   Treatment Facet   n Mean Median StdDev  Min  Max
#> 1  AB12345   AVAL       CRP Combination    BL 132 1.01   1.00   0.21 0.51 1.56
#> 2  AB12345   AVAL       CRP     Placebo    BL 134 1.00   1.00   0.20 0.50 1.58
#> 3  AB12345   AVAL       CRP    150mg QD    BL 134 0.97   0.99   0.24 0.47 1.63
#> 4  AB12345   AVAL       CRP       Comb.    BL 400 0.99   1.00   0.21 0.47 1.63
#> 5  AB12345   AVAL       CRP Combination   SCR 132 0.98   0.97   0.21 0.41 1.44
#> 6  AB12345   AVAL       CRP     Placebo   SCR 134 0.99   0.97   0.21 0.51 1.76
#> 7  AB12345   AVAL       CRP    150mg QD   SCR 134 0.99   1.00   0.22 0.40 1.47
#> 8  AB12345   AVAL       CRP       Comb.   SCR 400 0.99   0.98   0.22 0.40 1.76
#> 9  AB12345   AVAL       CRP Combination   W 1 132 1.01   1.02   0.23 0.44 1.53
#> 10 AB12345   AVAL       CRP     Placebo   W 1 134 1.01   1.03   0.22 0.42 1.43
#> 11 AB12345   AVAL       CRP    150mg QD   W 1 134 1.05   1.03   0.20 0.46 1.57
#> 12 AB12345   AVAL       CRP       Comb.   W 1 400 1.02   1.02   0.22 0.42 1.57
#> 13 AB12345   AVAL       CRP Combination   W 2 132 0.98   0.97   0.21 0.27 1.51
#> 14 AB12345   AVAL       CRP     Placebo   W 2 134 1.02   1.04   0.19 0.52 1.44
#> 15 AB12345   AVAL       CRP    150mg QD   W 2 134 0.99   1.01   0.21 0.48 1.48
#> 16 AB12345   AVAL       CRP       Comb.   W 2 400 1.00   1.01   0.20 0.27 1.51
#> 17 AB12345   AVAL       CRP Combination   W 3 132 0.99   0.99   0.19 0.50 1.57
#> 18 AB12345   AVAL       CRP     Placebo   W 3 134 0.98   0.97   0.19 0.52 1.44
#> 19 AB12345   AVAL       CRP    150mg QD   W 3 134 1.00   1.00   0.21 0.42 1.48
#> 20 AB12345   AVAL       CRP       Comb.   W 3 400 0.99   0.98   0.20 0.42 1.57
#> 21 AB12345   AVAL       CRP Combination   W 4 132 0.99   1.00   0.19 0.55 1.46
#> 22 AB12345   AVAL       CRP     Placebo   W 4 134 0.99   0.97   0.21 0.43 1.43
#> 23 AB12345   AVAL       CRP    150mg QD   W 4 134 1.04   1.06   0.20 0.54 1.53
#> 24 AB12345   AVAL       CRP       Comb.   W 4 400 1.01   1.01   0.20 0.43 1.53
#> 25 AB12345   AVAL       CRP Combination   W 5 132 0.98   0.98   0.19 0.55 1.59
#> 26 AB12345   AVAL       CRP     Placebo   W 5 134 1.02   1.01   0.19 0.58 1.51
#> 27 AB12345   AVAL       CRP    150mg QD   W 5 134 0.99   0.98   0.22 0.40 1.49
#> 28 AB12345   AVAL       CRP       Comb.   W 5 400 1.00   0.99   0.20 0.40 1.59
#>    PctMiss PctLOQ
#> 1        0    100
#> 2        0    100
#> 3        0    100
#> 4        0    100
#> 5        0    100
#> 6        0    100
#> 7        0    100
#> 8        0    100
#> 9        0    100
#> 10       0    100
#> 11       0    100
#> 12       0    100
#> 13       0    100
#> 14       0    100
#> 15       0    100
#> 16       0    100
#> 17       0    100
#> 18       0    100
#> 19       0    100
#> 20       0    100
#> 21       0    100
#> 22       0    100
#> 23       0    100
#> 24       0    100
#> 25       0    100
#> 26       0    100
#> 27       0    100
#> 28       0    100