Function to generate a boxplot Output rendered by teal.goshawk module g_boxplot returns boxplot visualization

A box plot is a method for graphically depicting groups of numerical data through their quartiles. Box plots may also have lines extending vertically from the boxes (whiskers) indicating variability outside the upper and lower quartiles, hence the term box-and-whisker. Outliers may be plotted as individual points. Box plots are non-parametric: they display variation in samples of a statistical population without making any assumptions of the underlying statistical distribution. The spacings between the different parts of the box indicate the degree of dispersion (spread) and skewness in the data, and show outliers. In addition to the points themselves, they allow one to visually estimate various L-estimators, notably the interquartile range, midhinge, range, mid-range, and trimean.

g_boxplot(data, biomarker, yaxis_var, trt_group, xaxis_var = NULL,
  loq_flag = "LOQFL", unit = NULL, timepoint = NULL,
  color_manual = NULL, shape_manual = NULL, box = TRUE,
  logscale = FALSE, ymax_scale = NULL, ymin_scale = NULL, dot_size = 2,
  alpha = 1, facet_ncol = NULL, hline = NULL, rotate_xlab = FALSE,
  font_size = NULL, armlabel = NULL, facet = NULL)

Arguments

data
data frame with variables which will be displayed in the plot.

biomarker
biomarker to visualize e.g. IGG.

yaxis_var
name of variable containing biomarker results displayed on Y-axis e.g. AVAL.

trt_group
name of variable representing treatment trt_group e.g. ARM.

xaxis_var
variable used to group the data on the x-axis.

loq_flag
name of variable containing LOQ flag e.g. LOQFL.

unit
biomarker unit label e.g. (U/L)

timepoint
text to include on the plot title

color_manual
vector of colour for trt_group

shape_manual
vector of shapes (used with log_flag)

box
add boxes to the plot (boolean)

logscale
use a log scale for the Y axis (boolean)

ymax_scale
maximum value for the Y axis

ymin_scale
minimum value for the Y axis

dot_size
plot dot size.

alpha
dot transparency (0 = transparent, 1 = opaque)

facet_ncol
number of facets per row. NULL = Use the default for facet_wrap

hline
y-axis value to position a horizontal line. NULL = No line

rotate_xlab
45 degree rotation of x-axis label values.

font_size
point size of tex to use. NULL is use default size

armlabel
label for the treatment symbols in the legend. If not specified then the label attribute for trt_group will be used. If there is no label attribute for trt_group, then the name of the parameter (in title case) will be used.

facet
variable to facet the plot by, or "None" if no faceting required.

Value

ggplot object

Examples

# NOT RUN {
library(random.cdisc.data)
library(dplyr)
library(goshawk)

# Example data for 100 patients from 3 treatment groups
ASL <- radam('ASL', start_with = list(
  STUDYID = "XX99999",
  USUBJID = paste0("XX99999-999-",1:100),
  ITTFL = 'Y',
  SEX = c("M", "F"),
  ARM = paste("ARM", LETTERS[1:3])
))

# Create an example ALB dataset from the 100 patient, with 6 visits with 3 
# parameters for each subject.  
ALB <- expand.grid(USUBJID = ASL$USUBJID
              , AVISIT = c("BASELINE",  paste0("VISIT ", 1:5))
              , PARAMCD = c("CRP", "IGA", "IGG")
              , stringsAsFactors = FALSE
    )

ALB <- ALB %>%
  mutate( AVAL = rnorm(nrow(ALB), mean = 50, sd = 8) )

ALB <- ALB %>%
  left_join(ALB %>%
              filter(AVISIT == "BASELINE") %>%
              select(USUBJID, PARAMCD, BASE = AVAL)
            , by = c("USUBJID", "PARAMCD")) %>%
  left_join(ASL, by = "USUBJID") %>%
  mutate(CHG = AVAL - BASE, PCHG = 100 * CHG/BASE) %>%
  mutate(LOQFL = ifelse(AVAL < 32, "Y", "N"))

# Example 1.

# Plot boxplots for IGA by Treatment (ARM) faceted by AVISIT, to display a
# boxplot for each visit.
# Specific colors are given for each treatment group.
# Values that are flagged by LOQFL are given a different symbol.
# Log scale for the Y axis.
# Points are plot with a transparency

# ALB is a VAD or ADAM format dataset, subsetted for the appropriate population.
# Select the parameter of interest to plot.
albplot <- ALB %>%
  filter(PARAMCD == "IGA")

g_boxplot(albplot
          , biomarker = "IGA"
          , yaxis_var = "AVAL"
          , trt_group = "ARM"
          , loq_flag = 'LOQFL'
          , timepoint = "over time"
          , unit = "U/L"
          #, color_manual = c('ARM A' = "#1F78B4", 'ARM B' = "#33A02C", 'ARM C' = "#601010")
          , shape_manual = c('N' = 1, 'Y' = 2, 'NA' = NULL)
          , hline = NULL
          , facet = "ARM"
          , xaxis_var = "AVISIT"
          , alpha = 0.5
          , logscale = FALSE
)

# Example 2
#
# A plot using mostly default values. 
albplot <- ALB %>%
  filter(PARAMCD == "CRP" & AVISIT == "BASELINE")

g_boxplot(albplot
          , biomarker = "CRP"
          , yaxis_var = "AVAL"
          , trt_group = "ARM"
          , xaxis_var = "AVISIT"
          , facet = "ARM"
          ,rotate_xlab = TRUE
)

# }